refactor(bot): update consumer imports for split track handler
updates all handler imports to use the new nowPlayingDisplay and
lastfmScrobbler modules. removes the original trackNowPlaying module.
