+++ packages/bot/src/handlers/eventHandler.ts
@@ -38,7 +38,7 @@ import { reactionRolesService } from '@lucky/shared/services'
 import { syncAllGuildFollowerRoles } from '../twitch/followerRoleSync'
 import { aiDevToolkitService } from '../services/AiDevToolkitService'
 import { namedSessionService } from '../services/musicManagement/namedSessions'
-import { cleanupGuildState } from './player/trackNowPlaying'
+import { cleanupGuildState } from './player/nowPlayingDisplay'
 import {
     recordGuildJoin,
     recordGuildLeave,
